+++ lib/api/users/delete-user.ts
@@ -1,14 +1,15 @@
 'use server';
 
+import { permissions } from '@/lib/permissions/permissions';
 import { api } from '@/lib/api/utils';
 import { revalidatePath } from 'next/cache';
 
 export default async function deleteUser(id: bigint) {
-  const result = await api.delete(`users/${id}`);
+  const result = await api.delete(`users/${id}`, permissions.users.delete);
 
-  if (!result.ok) {
-    throw Error('An error occurred when deleting the user');
+  if (result.ok) {
+    revalidatePath('/admin/users');
   }
 
-  revalidatePath('/admin/users');
+  return result;
 }
